+++ lib/tst_fs_has_free.c
@@ -27,7 +27,7 @@
 #include "tst_fs.h"
 
 int tst_fs_has_free_(void (*cleanup)(void), const char *path,
-		     unsigned int size, unsigned int mult)
+		     uint64_t size, unsigned int mult)
 {
 	struct statfs sf;
